escapealgo/requirements.txt
google-api-python-client==2.108.0
yt-dlp==2024.3.10
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.2
